Check permissions for modify the acl (#326)
* check permissions for add and remove member

* delete acl

* utils file

* check valid members and acls

* remove comment

* check updater of acl

* add check for delete and instert

* remove comment

* fix select *

* check add be a valid id

+++ packages/alea-frontend/pages/api/access-control/delete-acl.ts
@@ -3,16 +3,17 @@ import {
   checkIfPostOrSetError,
   executeTxnAndEndSet500OnError,
 } from '../comment-utils';
+import { isCurrentUserMemberOfAClupdater } from '../acl-utils/acl-common-utils';
 
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
   if (!checkIfPostOrSetError(req, res)) return;
-  // TODO: Check if from updaterACLId
   const id = req.body.id as string;
   if (!id || typeof id !== 'string') return res.status(422).send('Missing id.');
-
+  if (!await isCurrentUserMemberOfAClupdater(id, res, req))
+    return res.status(403).end();
   await executeTxnAndEndSet500OnError(
     res,
     'DELETE FROM AccessControlList WHERE id=?',
